+++ crypto/src/pqc/crypto/sphincsplus/SPHINCSPlusPublicKeyParameters.cs
@@ -38,7 +38,7 @@ namespace Org.BouncyCastle.Pqc.Crypto.SphincsPlus
 
         public byte[] GetEncoded()
         {
-            var id = Pack.UInt32_To_BE(SphincsPlusParameters.GetID(Parameters));
+            var id = Pack.UInt32_To_BE((uint)SphincsPlusParameters.GetID(Parameters));
             return Arrays.ConcatenateAll(id, m_pk.seed, m_pk.root);
         }
     }
